etl/simple_project/sqlite_to_postgres/requirements.txt
flake8==4.0.1
pre_commit==2.17
psycopg2==2.9.5
pytest==7.1.1
python-dotenv==0.20.0